리펙토링중_16
1. 강화로직 구현완료(UI연동, 범위와 속도 연동)

TODO
1. 몬스터 피격, 사망 애니메이션 동기화
2. 유닛 이동속도, 스폰간격, 투사체 속도 밸런싱
3. UI와 기능들 동기화(라이프, 남은 몬스터 수, 소지 금화)
4. 아군과 몬스터의 부모 클래스에 아군 로직들이 너무 많음 이거 고민 해 봐야 함
5. 데이터화 할 요소들 구분해놓기

+++ Assets/Refactoring/05_Datas/JUnitData.cs
@@ -14,10 +14,10 @@ public class JUnitData : ScriptableObject
 
     [Space(20)]
     [Header("스테이터스")]
-    public float AtkPower;
-    public float AtkRange;
-    public float AtkSpeed;
-    public float UpgradeCount;
+    public int AtkPower;
+    public int AtkRange;
+    public int AtkSpeed;
+    public int UpgradeCount;
 
     public JUnitData Clone()
     {
